requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.2
python-dotenv==0.20.0
openai==1.5.0
together==0.2.10
Flask==2.3.2